+++ packages/terra-slide-group/tests/wdio/slide-group-spec.js
@@ -11,32 +11,72 @@ Terra.describeViewports('Slide Group', ['medium'], () => {
   });
 
   describe('Non Animated', () => {
-    it('should advance the slide', () => {
+    it('should advance the slide with keyboard controls', () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/slide-group/non-animated-slide-group');
+
+      browser.keys(['Tab']);
+      $('#increment-1').isFocused();
+      Terra.validates.element('slide 1 non animated increment focused');
+
+      browser.keys(['Enter']);
+      Terra.validates.element('slide 2 non animated');
+    });
+
+    it('should go back a slide with keyboard controls and remain focused on originating element', () => {
+      browser.keys(['Tab', 'Tab']);
+      $('#decrement-2').isFocused();
+      Terra.validates.element('slide 2 non animated decrement focused');
+
+      browser.keys(['Enter']);
+      $('#increment-1').isFocused();
+      Terra.validates.element('slide 1 non animated increment focused');
+    });
+
+    it('should advance the slide with mouse controls', () => {
       $('#increment-1').click();
       Terra.validates.element('slide 2 non animated');
     });
 
-    it('should go back a slide', () => {
+    it('should go back a slide with mouse controls', () => {
       $('#decrement-2').click();
-      Terra.validates.element('slide 1 non animated');
+      Terra.validates.element('slide 1 non animated increment focused');
     });
   });
 
   describe('Animated', () => {
-    it('should advance the slide', () => {
-      browser.url('/raw/tests/cerner-terra-framework-docs/slide-group/animated-slide-group');
-      $('#increment-1').click();
+    it('should advance the slide with keyboard controls', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/slide-group/non-animated-slide-group');
 
+      browser.keys(['Tab']);
+      $('#increment-1').isFocused();
+      Terra.validates.element('slide 1 animated increment focused');
+
+      browser.keys(['Enter']);
       browser.pause(350);
       Terra.validates.element('slide 2 animated');
     });
 
-    it('should go back a slide', () => {
-      $('#decrement-2').click();
+    it('should go back a slide with keyboard controls and remain focused on originating element', () => {
+      browser.keys(['Tab', 'Tab']);
+      $('#decrement-2').isFocused();
+      Terra.validates.element('slide 2 animated decrement focused');
 
+      browser.keys(['Enter']);
       browser.pause(350);
-      Terra.validates.element('slide 1 animated');
+      $('#increment-1').isFocused();
+      Terra.validates.element('slide 1 animated increment focused');
+    });
+
+    it('should advance the slide with mouse controls', () => {
+      $('#increment-1').click();
+      browser.pause(350);
+      Terra.validates.element('slide 2 animated');
+    });
+
+    it('should go back a slide with mouse controls', () => {
+      $('#decrement-2').click();
+      browser.pause(350);
+      Terra.validates.element('slide 1 animated increment focused');
     });
   });
 });