+++ packages/terra-slide-panel-manager/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Update jest snapshot.
+
 ## 5.64.0 - (March 1, 2023)
 
 * Changed